+++ services/src/main/java/org/keycloak/protocol/oidc/TokenManager.java
@@ -813,20 +813,21 @@ public class TokenManager {
             //we could have multiple time this dynamic scope requested with different values
             //for multiple times requested this dynamic scope,  returned claim values consist all the requested values - if they exist
             //if a value is not containing in final value parameter -> remove this scope
-            if (finalToken.getOtherClaims().containsKey(cs.getName())) {
-                Object value = finalToken.getOtherClaims().get(cs.getName());
+            if (finalToken.getOtherClaims().containsKey(cs.getFilteredClaim() != null && !cs.getFilteredClaim().isEmpty() ? cs.getFilteredClaim() : cs.getName())) {
+                String filterClaim = cs.getFilteredClaim() != null && !cs.getFilteredClaim().isEmpty() ? cs.getFilteredClaim() : cs.getName();
+                Object value = finalToken.getOtherClaims().get(filterClaim);
                 List<String> requestedValues = scopeList.stream().filter(x -> x.contains(cs.getName()+":")).map(x -> x.replace(cs.getName()+":","")).collect(Collectors.toList());
                 if (requestedValues.size() >0 && value instanceof List<?>) {
                     //filter with all possible values
                     List<?> list = ((ArrayList<?>) value).stream().filter(x -> requestedValues.contains(x.toString())).collect(Collectors.toList());
                     if (list.isEmpty()) {
-                        finalToken.getOtherClaims().remove(cs.getName());
+                        finalToken.getOtherClaims().remove(filterClaim);
                     } else {
-                        finalToken.getOtherClaims().put(cs.getName(), list);
+                        finalToken.getOtherClaims().put(filterClaim, list);
                     }
                 } else  if (requestedValues.size() >0 ) {
                     if (!requestedValues.contains(value.toString()))
-                        finalToken.getOtherClaims().remove(cs.getName());
+                        finalToken.getOtherClaims().remove(filterClaim);
                 }
             }
         });